pyproject.toml
[project]
name = "chatbot-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.110.0",
    "pipreqs>=0.5.0",
    "asyncpg>=0.29.0,<0.30.0",
    "pyyaml==6.0.1",
    "uvicorn==0.28.0",
    "pydantic==2.7.0",
    "pydantic-settings==2.2.1",
    "pillow-heif==0.16.0",
    "bitsandbytes==0.43.1",
    "accelerate==0.29.2",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary==2.9.9",
    "starlette==0.36.3",
    "itsdangerous==2.2.0",
    "alembic>=1.13.0,<1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
